Issue 50 - extrai icones inline para components/icons/
Centraliza SVGs reutilizaveis em app/components/icons/ pra acabar com
duplicacao (ErrorIcon estava em 3 lugares) e dar um lugar canonico
para o crescimento da biblioteca de icones.

Novos arquivos:
- BrandMark, ErrorIcon, CheckIcon, ChevronDownIcon, ArrowRightIcon,
  PlusIcon (um arquivo cada)
- BadgeStatusIcon (funcao com switch dos 9 status)
- DashboardCardIcon (funcao com switch dos 4 cards)
- index.ts com barrel exports

Consumidores atualizados:
- Sidebar usa BrandMark
- Badge usa BadgeStatusIcon
- Input usa ErrorIcon e CheckIcon
- Textarea usa ErrorIcon
- Select usa ChevronDownIcon, CheckIcon e ErrorIcon
- dashboard/page usa DashboardCardIcon
- ComponentsShowcase usa ArrowRightIcon e PlusIcon

Button: Spinner inline removido. Consolida com ui/Spinner.tsx, que
ganha prop opcional className pra permitir customizacao de tamanho e
velocidade da animacao sem quebrar os tres callers existentes.
