+++ frontend/app/components/icons/ErrorIcon.tsx
@@ -1,0 +1,21 @@
+type ErrorIconProps = {
+  className?: string;
+};
+
+export function ErrorIcon({ className = "w-3 h-3 shrink-0" }: ErrorIconProps) {
+  return (
+    <svg
+      className={className}
+      viewBox="0 0 16 16"
+      fill="none"
+      stroke="currentColor"
+      strokeWidth={1.75}
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      aria-hidden="true"
+    >
+      <circle cx="8" cy="8" r="6.5" />
+      <path d="M8 5v3.5M8 11v.5" />
+    </svg>
+  );
+}
