+++ frontend/app/components/icons/PlusIcon.tsx
@@ -1,0 +1,20 @@
+type PlusIconProps = {
+  className?: string;
+};
+
+export function PlusIcon({ className = "w-3.5 h-3.5" }: PlusIconProps) {
+  return (
+    <svg
+      className={className}
+      viewBox="0 0 16 16"
+      fill="none"
+      stroke="currentColor"
+      strokeWidth={1.75}
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      aria-hidden="true"
+    >
+      <path d="M8 3v10M3 8h10" />
+    </svg>
+  );
+}
