+++ frontend/app/components/ui/Spinner.tsx
@@ -1,7 +1,13 @@
-export default function Spinner() {
+type SpinnerProps = {
+  className?: string;
+};
+
+export default function Spinner({
+  className = "animate-spin h-4 w-4",
+}: SpinnerProps) {
   return (
     <svg
-      className="animate-spin h-4 w-4"
+      className={className}
       xmlns="http://www.w3.org/2000/svg"
       fill="none"
       viewBox="0 0 24 24"
@@ -9,7 +15,9 @@ export default function Spinner() {
     >
       <circle
         className="opacity-25"
-        cx="12" cy="12" r="10"
+        cx="12"
+        cy="12"
+        r="10"
         stroke="currentColor"
         strokeWidth="4"
       />
@@ -20,4 +28,4 @@ export default function Spinner() {
       />
     </svg>
   );
-};
+}
