+++ frontend/app/styleguide/ComponentsShowcase.tsx
@@ -6,6 +6,7 @@ import { Button } from "../components/ui/Button/Button";
 import { Input } from "../components/ui/Input/Input";
 import { Select } from "../components/ui/Select/Select";
 import { Textarea } from "../components/ui/Textarea/Textarea";
+import { ArrowRightIcon, PlusIcon } from "../components/icons";
 
 const BADGE_STATUSES: BadgeStatus[] = [
   "planejado",
@@ -40,22 +41,6 @@ function Group({ title, children }: { title: string; children: React.ReactNode }
         {children}
       </div>
     </div>
-  );
-}
-
-function ArrowRight() {
-  return (
-    <svg viewBox="0 0 16 16" width={14} height={14} fill="none" stroke="currentColor" strokeWidth={1.75} strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
-      <path d="M3 8h10M9 4l4 4-4 4" />
-    </svg>
-  );
-}
-
-function PlusIcon() {
-  return (
-    <svg viewBox="0 0 16 16" width={14} height={14} fill="none" stroke="currentColor" strokeWidth={1.75} strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
-      <path d="M8 3v10M3 8h10" />
-    </svg>
   );
 }
 
@@ -112,8 +97,8 @@ export function ComponentsShowcase() {
 
           <Group title="leftIcon / rightIcon">
             <Button leftIcon={<PlusIcon />}>Novo</Button>
-            <Button variant="secondary" rightIcon={<ArrowRight />}>Próximo</Button>
-            <Button variant="ghost" leftIcon={<PlusIcon />} rightIcon={<ArrowRight />}>
+            <Button variant="secondary" rightIcon={<ArrowRightIcon />}>Próximo</Button>
+            <Button variant="ghost" leftIcon={<PlusIcon />} rightIcon={<ArrowRightIcon />}>
               Ambos
             </Button>
           </Group>
